+++ react/FormattedCurrency.tsx
@@ -51,7 +51,8 @@ interface Props {
 }
 
 function FormattedCurrency({ value, classes }: Props) {
-  const { culture } = useRuntime()
+  const { culture, getSettings } = useRuntime()
+  const settings = getSettings('vtex.store')
   const intl = useIntl()
   const { handles } = useCssHandles(CSS_HANDLES, { classes })
 
@@ -97,6 +98,16 @@ function FormattedCurrency({ value, classes }: Props) {
             lei
           </span>
         )
+      }
+
+      if (part.type === 'currency' && 
+      culture?.customCurrencySymbol.length > 0 &&
+      settings?.enableCustomCurrencySymbol) {
+        return (
+            <span key={index} className={handle}>
+              {culture.customCurrencySymbol}
+            </span>
+          )
       }
 
       return (