+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Custom Currency Symbol from Trade Policy into `FormattedCurrency` when available.
 
 ## [0.3.0] - 2020-12-17
 ### Changed